+++ scripts/macros/2024/spells/bless.js
@@ -1,0 +1,7 @@
+import {bless as blessLegacy} from '../../spells/bless.js';
+export let bless = {
+    name: 'Bless',
+    version: '1.1.17',
+    rules: 'modern',
+    midi: blessLegacy.midi
+};